minimal stream client shuffling
